+++ src/Checks/GlobalCheck.hs
@@ -12,6 +12,7 @@ import qualified Data.Text as T
 import Optics
 
 import Types
+import Util.ExpandVariables
 import Util.RunCommand
 
 -- Implements SPEC-1 @relation(SPEC-1, scope=file)
@@ -20,7 +21,9 @@ runGlobalCheck :: GlobalCheck -> ReaderT CheckContext IO CheckReport
 runGlobalCheck check = do
     wd <- asks (view #directory)
     commit <- asks (view #commit)
-    let cmd = check.command & T.unpack
+    vars <- asks (view #variables)
+    let cmdRaw = check.command
+        cmd = T.unpack . expandVariables vars $ cmdRaw
         expected = check.expectedExit
     (exit, out) <- runCommandWithStderrIn wd cmd
     let result =