+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to the
 - Changed default commit range from `HEAD` to `HEAD~1..HEAD`.
 - Changed return code to failure if any checks failed during the run.
 - Changed config file resolution to be relative to repository under check.
+- Added expansion of variables `${filename}` and `${commit-range}` in check commands
 
 ## 0.1.0 - 2025-10-07